requirements.txt
# Core
pendulum>=2.1.2

# Airflow providers
apache-airflow-providers-common-sql>=1.9.0

# Database
psycopg2-binary>=2.9.9

# Networking
requests>=2.31.0

# Environment management
python-dotenv>=1.0.1

# Testing
pytest>=8.0.0